+++ packages/parser/src/parser/index.ts
@@ -1,5 +1,6 @@
 export * from './condition-parser.js';
 export * from './directive-parser.js';
+export * from './heading-sections.js';
 export * from './parse-file.js';
 export * from './rank-parser.js';
 export * from './reference-resolver.js';